Refactor the Comment class for reusability
