+++ src/components/CommentContent.css
@@ -1,0 +1,3 @@
+.CommentContent {
+	padding: 14px;
+}
