+++ src/components/CommentMeta.css
@@ -1,0 +1,26 @@
+.CommentMeta {
+	display: flex;
+	justify-content: space-between;
+	background: #e5f5fa;
+	color: #82878c;
+	padding: 0.7rem 14px;
+}
+
+.CommentMeta p {
+	margin: 0;
+}
+
+.CommentMeta-author {
+	color: #23282d;
+	font-weight: bold;
+}
+
+.CommentMeta-author-label {
+	display: inline-block;
+	font-size: 0.9em;
+	border: 1px solid #bfe7f3;
+	background: rgba( 255, 255, 255, 0.1 );
+	padding: 0.15em 0.4em;
+	margin-top: -0.15em;
+	margin-right: 0.6em;
+}
